Guard against empty sanitizeEnvironment setting
This code was causing a CI failure, probably due to this setting not
being populated by Atom APIs when running in CI.  It also hangs Atom
completely when deserializing a TerminalView at Atom startup.  To
repro, create a terminal tab and reload the Atom window while the tab
is still open.